refactoring: move `workloadinstances` package
