+++ pilot/pkg/serviceregistry/util/workloadinstances/index.go
