+++ pilot/pkg/serviceregistry/util/workloadinstances/map.go
